* The minimum version of Go in `ydb-go-sdk` has been raised to `go1.20`
