+++ internal/xsql/badconn/badconn_test.go
@@ -1,6 +1,3 @@
-//go:build go1.18
-// +build go1.18
-
 package badconn
 
 import (
