+++ internal/xsql/unwrap.go
@@ -1,10 +1,6 @@
-//go:build !go1.18
-// +build !go1.18
-
 package xsql
 
 import (
-	"context"
 	"database/sql"
 	"fmt"
 
@@ -12,19 +8,26 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
 )
 
-func Unwrap(db *sql.DB) (connector *Connector, err error) {
-	c, err := db.Conn(context.Background())
-	if err != nil {
-		return nil, badconn.Map(xerrors.WithStackTrace(err))
-	}
-	if err = c.Raw(func(driverConn interface{}) error {
-		if cc, ok := driverConn.(*conn); ok {
-			connector = cc.connector
-			return nil
+func Unwrap[T *sql.DB | *sql.Conn](v T) (connector *Connector, err error) {
+	switch vv := any(v).(type) {
+	case *sql.DB:
+		d := vv.Driver()
+		if dw, ok := d.(*driverWrapper); ok {
+			return dw.c, nil
 		}
-		return xerrors.WithStackTrace(badconn.Map(fmt.Errorf("%+v is not a *conn", driverConn)))
-	}); err != nil {
-		return nil, badconn.Map(xerrors.WithStackTrace(err))
+		return nil, xerrors.WithStackTrace(fmt.Errorf("%T is not a *driverWrapper", d))
+	case *sql.Conn:
+		if err = vv.Raw(func(driverConn interface{}) error {
+			if cc, ok := driverConn.(*conn); ok {
+				connector = cc.connector
+				return nil
+			}
+			return xerrors.WithStackTrace(fmt.Errorf("%T is not a *conn", driverConn))
+		}); err != nil {
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
+		}
+		return connector, nil
+	default:
+		return nil, xerrors.WithStackTrace(fmt.Errorf("unknown type %T for Unwrap", vv))
 	}
-	return connector, nil
 }
