--- retry/errors_go1.18.go
@@ -1,17 +1,0 @@
-//go:build go1.18
-// +build go1.18
-
-package retry
-
-import (
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
-)
-
-func unwrapErrBadConn(err error) error {
-	var e *badconn.Error
-	if xerrors.As(err, &e) {
-		return e.Origin()
-	}
-	return err
-}
