+++ sql_unwrap.go
@@ -1,6 +1,3 @@
-//go:build !go1.18
-// +build !go1.18
-
 package ydb
 
 import (
@@ -10,8 +7,8 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql"
 )
 
-func Unwrap(db *sql.DB) (*Driver, error) {
-	c, err := xsql.Unwrap(db)
+func Unwrap[T *sql.DB | *sql.Conn](v T) (*Driver, error) {
+	c, err := xsql.Unwrap(v)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
