--- sql_unwrap_go1.18.go
@@ -1,22 +1,0 @@
-//go:build go1.18
-// +build go1.18
-
-package ydb
-
-import (
-	"database/sql"
-
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql"
-)
-
-func Unwrap[T *sql.DB | *sql.Conn](v T) (*Driver, error) {
-	c, err := xsql.Unwrap(v)
-	if err != nil {
-		return nil, xerrors.WithStackTrace(err)
-	}
-	d.connectorsMtx.RLock()
-	defer d.connectorsMtx.RUnlock()
-
-	return d.connectors[c], nil
-}
